Polyfill shapeless instead of extracting its usages
