+++ core/src/main/scala-2/io/chrisdavenport/epimetheus/ShapelessPolyfill.scala
@@ -1,0 +1,9 @@
+package io.chrisdavenport.epimetheus
+
+trait ShapelessPolyfill {
+
+  type Nat = shapeless.Nat
+
+  type Sized[+Repr,L <: Nat] = shapeless.Sized[Repr, L]
+
+}
