+++ core/src/main/scala/io/chrisdavenport/epimetheus/Counter.scala
@@ -61,7 +61,7 @@ sealed abstract class Counter[F[_]]{
 /**
  * Counter Constructors, and Unsafe Counter Access
  */
-object Counter extends LabelledCounters {
+object Counter extends ShapelessPolyfill {
 
   /**
    * Constructor for a Counter with no labels.
@@ -75,6 +75,32 @@ object Counter extends LabelledCounters {
     out <- Sync[F].delay(c.register(CollectorRegistry.Unsafe.asJava(cr)))
   } yield new NoLabelsCounter[F](out)
 
+  /**
+   * Constructor for a labelled [[Counter]].
+   *
+   * This generates a specific number of labels via `Sized`, in combination with a function
+   * to generate an equally `Sized` set of labels from some type. Values are applied by position.
+   *
+   * This counter needs to have a label applied to the [[UnlabelledCounter]] in order to
+   * be measureable or recorded.
+   *
+   * @param cr CollectorRegistry this [[Counter]] will be registred with
+   * @param name The name of the Counter -  By convention, the names of Counters are suffixed by `_total`.
+   * @param help The help string of the metric
+   * @param labels The name of the labels to be applied to this metric
+   * @param f Function to take some value provided in the future to generate an equally sized list
+   *  of strings as the list of labels. These are assigned to labels by position.
+   */
+  def labelled[F[_]: Sync, A, N <: Nat](
+    cr: CollectorRegistry[F],
+    name: Name,
+    help: String,
+    labels: Sized[IndexedSeq[Label], N],
+    f: A => Sized[IndexedSeq[String], N]
+  ): F[UnlabelledCounter[F, A]] = for {
+    c <- Sync[F].delay(JCounter.build().name(name.getName).help(help).labelNames(labels.unsized.map(_.getLabel):_*))
+    out <- Sync[F].delay(c.register(CollectorRegistry.Unsafe.asJava(cr)))
+  } yield new UnlabelledCounterImpl[F, A](out, f.andThen(_.unsized))
 
   private final class NoLabelsCounter[F[_]: Sync] private[Counter] (private[Counter] val underlying: JCounter) extends Counter[F] {
     override def get: F[Double] = Sync[F].delay(underlying.get)
@@ -107,7 +133,7 @@ object Counter extends LabelledCounters {
     def mapK[G[_]](fk: F ~> G): UnlabelledCounter[G, A] = new MapKUnlabelledCounter[F, G, A](this, fk)
   }
 
-  private[epimetheus] final class UnlabelledCounterImpl[F[_]: Sync, A] private[epimetheus](
+  private final class UnlabelledCounterImpl[F[_]: Sync, A] private[Counter](
     private[Counter] val underlying: JCounter,
     private val f: A => IndexedSeq[String]
   ) extends UnlabelledCounter[F, A]{
