+++ core/src/main/scala/io/chrisdavenport/epimetheus/Gauge.scala
@@ -75,7 +75,7 @@ sealed abstract class Gauge[F[_]]{
 /**
  * Gauge Constructors, and Unsafe Gauge Access
  */
-object Gauge extends LabelledGauges {
+object Gauge extends ShapelessPolyfill {
 
   // Convenience
   def incIn[F[_], A](g: Gauge[F], fa: F[A])(implicit C: MonadCancel[F, _]): F[A] =
@@ -103,6 +103,34 @@ object Gauge extends LabelledGauges {
     c <- Sync[F].delay(JGauge.build().name(name.getName).help(help))
     out <- Sync[F].delay(c.register(CollectorRegistry.Unsafe.asJava(cr)))
   } yield new NoLabelsGauge[F](out)
+
+  /**
+   * Constructor for a labelled [[Gauge]].
+   *
+   * This generates a specific number of labels via `Sized`, in combination with a function
+   * to generate an equally `Sized` set of labels from some type. Values are applied by position.
+   *
+   * This counter needs to have a label applied to the [[UnlabelledGauge]] in order to
+   * be measureable or recorded.
+   *
+   * @param cr CollectorRegistry this [[Gauge]] will be registred with
+   * @param name The name of the [[Gauge]].
+   * @param help The help string of the metric
+   * @param labels The name of the labels to be applied to this metric
+   * @param f Function to take some value provided in the future to generate an equally sized list
+   *  of strings as the list of labels. These are assigned to labels by position.
+   */
+  def labelled[F[_]: Sync, A, N <: Nat](
+    cr: CollectorRegistry[F],
+    name: Name,
+    help: String,
+    labels: Sized[IndexedSeq[Label], N],
+    f: A => Sized[IndexedSeq[String], N]
+  ): F[UnlabelledGauge[F, A]] = for {
+    c <- Sync[F].delay(JGauge.build().name(name.getName).help(help).labelNames(labels.unsized.map(_.getLabel):_*))
+    out <- Sync[F].delay(c.register(CollectorRegistry.Unsafe.asJava(cr)))
+  } yield new UnlabelledGaugeImpl[F, A](out, f.andThen(_.unsized))
+
 
   private final class NoLabelsGauge[F[_]: Sync] private[Gauge] (
     private[Gauge] val underlying: JGauge
